fix: share network log sanitization
