+++ turbo/apps/api/src/signals/routes/__tests__/run-reads.bdd.test.ts
@@ -1437,7 +1437,9 @@ describe("RUN-04: agent run telemetry families", () => {
             latency_ms: 0,
             request_size: null,
             response_size: null,
+            request_body_encoding: "weird",
             firewall_params: null,
+            connector_diagnostic_env_names: ["FAL_TOKEN", 7],
             auth_resolved_secrets: null,
             error: null,
           },
